# docs/openapi.yaml
basePath: /api
...
host: localhost:8090
info:
  contact: {}
  description: Self-hosted CTF platform API. Most write endpoints require authentication
    via JWT cookie (auth_token).
  title: hCTF API
  version: 1.0.0
paths:
...
  /hints/{id}/unlock:
    post:
      parameters:
      - description: Hint ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: 'Empty response; triggers HX-Trigger: refreshHints'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Unlock a hint for a question (costs points)
      tags:
      - Hints
...
  /questions/{questionId}/hints:
    get:
      description: Returns hints with unlock status. Unlocked hint content is visible.
        Locked hints show cost and unlock button.
      parameters:
      - description: Question ID
        in: path
        name: questionId
        required: true
        type: string
      produces:
      - text/html
      responses:
        "200":
          description: HTML fragments with hint cards
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get hints for a question as HTML fragments (for HTMX)
      tags:
      - Hints
...
  /sql/snapshot:
    get:
      description: Returns the current state of the SQL playground database for use
        with DuckDB WASM.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - CookieAuth: []
      summary: Get a snapshot of the SQL playground database
      tags:
      - SQL
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"